feat: refactor model initialization, checkpoint loading, and path derivation with FSDP2 integration
- Add use_meta_init and lora_path params to Actor for direct LoRA adapter loading
- Simplify batch_inference.py and interactive_chat.py to use non meta-init path
- Clean up checkpoint.py: remove try/except fallback, move import to top level
- Update all example scripts: --save_path → --ckpt_save_path, remove invalid --ckpt_path,
  fix --vllm_enable_is_correction → --enable_vllm_is_correction
- Remove unused CLI args (--lora_rank/alpha/dropout/target_modules/device_map) from inference scripts

+++ openrlhf/models/loss.py
@@ -47,7 +47,7 @@ class GPTLMLoss(nn.Module):
             log_probs = log_probs_from_logits(shift_logits, shift_labels)
             mask = shift_labels != self.IGNORE_INDEX
             return -(log_probs * mask).sum() / mask.sum()
-        return self.loss(shift_logits.view(-1, shift_logits.size(-1)), shift_labels.view(-1))
+        return self.loss(shift_logits.float().view(-1, shift_logits.size(-1)), shift_labels.view(-1))
 
 
 class SFTLoss(nn.Module):
@@ -442,7 +442,7 @@ class PRMLoss(nn.Module):
             logits = logits[placeholder_mask].squeeze(1)
             labels = labels_ph
 
-        loss = self.loss(logits, labels)
+        loss = self.loss(logits.float(), labels)
         if not return_acc:
             return loss
